Roundtrip all SQL rows through our WIT SQL model (#162)
* Roundtrip all SQL rows through our WIT SQL model

This is a small first step to using the WIT SQL model everywhere.

Unfortunately the underlying API we're using doesn't give us the
type information we need to do this correctly, so this relies on
reflection to extract data we don't otherwise have access to. I've
structured this reflection hack to fail loudly if the implementation
details change in a future release of the Vert.x Postgresql client.

* Track column types

+++ os/server/db/src/jvmMain/kotlin/com/wasmo/db/schemaversion/SchemaVersionQueries.kt
@@ -48,8 +48,8 @@ private suspend fun ensureSchemaVersionStored(versionZeroMigrationHash: ByteStri
     // https://www.postgresql.org/docs/current/sql-prepare.html
     """
     CREATE TABLE IF NOT EXISTS DatabaseSchemaVersion (
-      id INTEGER NOT NULL PRIMARY KEY DEFAULT $DB_SCHEMA_VERSION_ID,
-      version INTEGER NOT NULL,
+      id BIGINT NOT NULL PRIMARY KEY DEFAULT $DB_SCHEMA_VERSION_ID,
+      version BIGINT NOT NULL,
       -- hash of the Map of migration SQLs up to and including version.
       migration_hash BYTEA NOT NULL
     )
